+++ pybop/_optimisation.py
@@ -1,16 +1,6 @@
 import warnings
 
-import pybop
-
-DEFAULT_OPTIMISER_OPTIONS = dict(
-    x0=None,
-    bounds=None,
-    sigma0=0.1,
-    verbose=False,
-    physical_viability=True,
-    allow_infeasible_solutions=True,
-    _max_iterations=None,
-)
+from pybop import BaseCost
 
 
 class Optimisation:
@@ -19,21 +9,15 @@ class Optimisation:
 
     This class serves as a base class for creating optimisers. It provides a basic structure for
     an optimisation algorithm, including the initial setup and a method stub for performing the
-    optimisation process. Child classes should override set_options and the _run method with a
-    specific algorithm.
+    optimisation process. Child classes should override _set_up_optimiser and the _run method with
+    a specific algorithm.
 
     Parameters
     ----------
     cost : pybop.BaseCost or pints.ErrorMeasure
         An objective function to be optimised, which can be either a pybop.Cost or PINTS error measure
-    sigma0 : float or sequence, optional
-        Initial step size or standard deviation for the optimiser (default: None).
-    verbose : bool, optional
-        If True, the optimisation progress is printed (default: False).
-    physical_viability : bool, optional
-        If True, the feasibility of the optimised parameters is checked (default: True).
-    allow_infeasible_solutions : bool, optional
-        If True, infeasible parameter values will be allowed in the optimisation (default: True).
+    **optimiser_kwargs : optional
+            Valid option keys and their values.
 
     Attributes
     ----------
@@ -41,12 +25,16 @@ class Optimisation:
         Initial parameter values for the optimisation.
     bounds : dict
         Dictionary containing the parameter bounds with keys 'lower' and 'upper'.
-    _n_parameters : int
-        Number of parameters in the optimisation problem.
     sigma0 : float or sequence
         Initial step size or standard deviation for the optimiser.
+    verbose : bool, optional
+        If True, the optimisation progress is printed (default: False).
+    physical_viability : bool, optional
+        If True, the feasibility of the optimised parameters is checked (default: True).
+    allow_infeasible_solutions : bool, optional
+        If True, infeasible parameter values will be allowed in the optimisation (default: True).
     log : list
-        Log of the optimisation process.
+        A log of the parameter values tried during the optimisation.
     """
 
     def __init__(
@@ -54,79 +42,73 @@ class Optimisation:
         cost,
         **optimiser_kwargs,
     ):
-        self.cost = cost
-        self.__dict__.update(DEFAULT_OPTIMISER_OPTIONS)
-        if isinstance(cost, pybop.BaseCost):
+        # First set attributes to default values
+        self.x0 = None
+        self.bounds = None
+        self.sigma0 = 0.1
+        self.verbose = False
+        self.log = []
+        self._minimising = True
+        self.physical_viability = False
+        self.allow_infeasible_solutions = False
+        self.default_max_iterations = 1000
+        self.result = None
+
+        if isinstance(cost, BaseCost):
+            self.cost = cost
             self.x0 = cost.x0
             self.bounds = cost.bounds
             self.sigma0 = cost.sigma0
-            self._n_parameters = cost._n_parameters
-        self.log = []
-        self.set_max_iterations()
-        self.set_allow_infeasible_solutions()
-        self.set_options(**optimiser_kwargs)
+            self._minimising = cost._minimising
+            self.set_allow_infeasible_solutions()
+        else:
+            warnings.warn(
+                "The cost is not an instance of pybop.BaseCost. Continuing "
+                + "under the assumption that it is a callable function.",
+                UserWarning,
+            )
+            self.cost = BaseCost()
 
-        # Check if minimising or maximising
-        if isinstance(self.cost, pybop.BaseLikelihood):
-            self.cost._minimising = False
-        self._minimising = self.cost._minimising
+            def cost_evaluate(x, grad=None):
+                return cost(x)
 
-    def set_options(self, **optimiser_kwargs):
+            self.cost._evaluate = cost_evaluate
+
+        self.unset_options = optimiser_kwargs
+        self.set_base_options()
+        self._set_up_optimiser()
+
+        # Throw an error if any options remain
+        if self.unset_options:
+            raise ValueError(f"Unrecognised keyword arguments: {self.unset_options}")
+
+    def set_base_options(self):
         """
-        Update the optimiser options and check that all have been applied.
-
-        Parameters
-        ----------
-        **optimiser_kwargs : optional
-            Valid option keys and their values.
+        Update the base optimiser options and remove them from the options dictionary.
         """
-        # Update and remove optimiser options from the optimiser_kwargs dictionary
-        # in the child class first and then in this base class
-        optimiser_kwargs = self._set_options(**optimiser_kwargs)
-
-        key_list = list(optimiser_kwargs.keys())
+        key_list = list(self.unset_options.keys())
         for key in key_list:
             if key in ["x0", "bounds", "sigma0", "verbose"]:
-                self.__dict__.update({key: optimiser_kwargs.pop(key)})
+                self.__dict__.update({key: self.unset_options.pop(key)})
             elif key == "allow_infeasible_solutions":
-                self.allow_infeasible_solutions = self.set_allow_infeasible_solutions(
-                    optimiser_kwargs.pop(key)
-                )
-            elif key == "parallel":
-                self.set_parallel(optimiser_kwargs.pop(key))
+                self.set_allow_infeasible_solutions(self.unset_options.pop(key))
 
-        # Throw an error if any arguments remain
-        if optimiser_kwargs:
-            raise ValueError(f"Unrecognised keyword arguments: {optimiser_kwargs}")
-
-    def _set_options(self, **optimiser_kwargs):
+    def _set_up_optimiser(self):
         """
-        Update the optimiser options and remove the corresponding entries from the
-        optimiser_kwargs dictionary in advance of passing to the parent class
-        - this function is to be overwritten by child classes.
+        Parse optimiser options and prepare the optimiser.
 
-        Parameters
-        ----------
-        **optimiser_kwargs : optional
-            Valid option keys and their values.
+        This method should be implemented by child classes.
 
-        Returns
-        -------
-        optimiser_kwargs : dict
-            Remaining option keys and their values.
+        Raises
+        ------
+        NotImplementedError
+            If the method has not been implemented by the subclass.
         """
-        return optimiser_kwargs
+        raise NotImplementedError
 
-    def run(self, **optimiser_kwargs):
+    def run(self):
         """
         Run the optimisation and return the optimised parameters and final cost.
-
-        Parameters
-        ----------
-        **optimiser_kwargs : optional
-            Valid option keys and their values, for example:
-            x0 : ndarray
-                Initial guess for the parameters.
 
         Returns
         -------
@@ -135,12 +117,10 @@ class Optimisation:
         final_cost : float
             The final cost associated with the best parameters.
         """
-        self.set_options(**optimiser_kwargs)
-
         x, final_cost = self._run()
 
         # Store the optimised parameters
-        if self.cost.problem is not None:
+        if hasattr(self.cost, "parameters"):
             self.store_optimised_parameters(x)
 
         # Store the log
@@ -207,23 +187,6 @@ class Optimisation:
         """
         return "Optimisation"
 
-    def set_max_iterations(self, iterations=1000):
-        """
-        Set the maximum number of iterations as a stopping criterion.
-        Credit: PINTS
-
-        Parameters
-        ----------
-        iterations : int, optional
-            The maximum number of iterations to run.
-            Set to `None` to remove this stopping criterion.
-        """
-        if iterations is not None:
-            iterations = int(iterations)
-            if iterations < 0:
-                raise ValueError("Maximum number of iterations cannot be negative.")
-        self._max_iterations = iterations
-
     def set_allow_infeasible_solutions(self, allow=True):
         """
         Set whether to allow infeasible solutions or not.
@@ -237,7 +200,7 @@ class Optimisation:
         self.physical_viability = allow
         self.allow_infeasible_solutions = allow
 
-        if self.cost.problem is not None and hasattr(self.cost.problem, "_model"):
+        if hasattr(self.cost, "problem") and hasattr(self.cost.problem, "_model"):
             self.cost.problem._model.allow_infeasible_solutions = (
                 self.allow_infeasible_solutions
             )