+++ pybop/plotting/plot_convergence.py
@@ -30,19 +30,25 @@ def plot_convergence(optim, show=True, **layout_kwargs):
     cost = optim.cost
     log = optim.log
 
-    # Compute the minimum cost for each iteration
-    min_cost_per_iteration = [
-        min((cost(solution) for solution in log_entry), default=np.inf)
-        for log_entry in log
-    ]
+    # Find the best cost from each iteration
+    if optim._minimising:
+        best_cost_per_iteration = [
+            min((cost(solution) for solution in log_entry), default=np.inf)
+            for log_entry in log
+        ]
+    else:
+        best_cost_per_iteration = [
+            max((cost(solution) for solution in log_entry), default=-np.inf)
+            for log_entry in log
+        ]
 
     # Generate a list of iteration numbers
-    iteration_numbers = list(range(1, len(min_cost_per_iteration) + 1))
+    iteration_numbers = list(range(1, len(best_cost_per_iteration) + 1))
 
     # Create a plotting dictionary
     plot_dict = pybop.StandardPlot(
         x=iteration_numbers,
-        y=min_cost_per_iteration,
+        y=best_cost_per_iteration,
         layout_options=dict(
             xaxis_title="Iteration", yaxis_title="Cost", title="Convergence"
         ),