+++ tests/unit/test_optimisation.py
@@ -92,16 +92,12 @@ class TestOptimisation:
         assert optim.cost is not None
         assert optim.name() == expected_name
 
-        # Test construction without bounds
-        bounds = cost.bounds
-        cost.bounds = None
-        optim = optimiser(cost=cost)
-        assert optim.bounds is None
-        if issubclass(optimiser, pybop.BasePintsOptimiser):
-            assert optim._boundaries is None
-
-        # Reset
-        cost.bounds = bounds
+        if optimiser not in [pybop.SciPyDifferentialEvolution]:
+            # Test construction without bounds
+            optim = optimiser(cost=cost, bounds=None)
+            assert optim.bounds is None
+            if issubclass(optimiser, pybop.BasePintsOptimiser):
+                assert optim._boundaries is None
 
     @pytest.mark.parametrize(
         "optimiser",
@@ -121,17 +117,13 @@ class TestOptimisation:
     def test_optimiser_kwargs(self, cost, optimiser):
         optim = optimiser(cost=cost, maxiter=1)
 
-        # Check and update maximum iterations
+        # Check maximum iterations
         optim.run()
-        assert optim._max_iterations == 1
         assert optim._iterations == 1
-        optim.run(max_iterations=10)
-        assert optim._max_iterations == 10
-        assert optim._iterations <= 10
 
         if issubclass(optimiser, (pybop.GradientDescent, pybop.Adam, pybop.NelderMead)):
-            # Unused bounds
-            optim.run(bounds=cost.bounds)
+            # Ignored bounds
+            optim = optimiser(cost=cost, bounds=cost.bounds)
             assert optim.bounds is None
         elif issubclass(optimiser, pybop.PSO):
             assert optim.bounds == cost.bounds
@@ -141,18 +133,17 @@ class TestOptimisation:
                 ValueError,
                 match="Either all bounds or no bounds must be set",
             ):
-                optim.run(bounds=bounds)
-            # Reset
-            optim.set_options(bounds=cost.bounds)
+                optim = optimiser(cost=cost, bounds=bounds)
         else:
             # Check and update bounds
             assert optim.bounds == cost.bounds
             bounds = {"upper": [0.63], "lower": [0.57]}
-            optim.run(bounds=bounds)
+            optim = optimiser(cost=cost, bounds=bounds)
             assert optim.bounds == bounds
 
         if issubclass(optimiser, pybop.BasePintsOptimiser):
-            optim.run(
+            optim = optimiser(
+                cost=cost,
                 use_f_guessed=True,
                 parallel=False,
                 min_iterations=3,
@@ -170,52 +161,43 @@ class TestOptimisation:
             bounds = [
                 (lower, upper) for lower, upper in zip(bounds["lower"], bounds["upper"])
             ]
-            optim.run(bounds=bounds, tol=1e-2)
+            optim = optimiser(cost=cost, bounds=bounds, tol=1e-2)
             assert optim.bounds == bounds
-            assert optim.tol == 1e-2
 
             # Pass nested options
-            options = dict(maxiter=10)
-            optim.run(options=options)
-            optim.run(maxiter=5, options=options)
-            assert optim._max_iterations == 5
-
-            # Trigger unused input warning
-            optim.set_max_unchanged_iterations(5)
+            optim = optimiser(cost=cost, options=dict(maxiter=10))
+            with pytest.raises(
+                Exception,
+                match="A duplicate maxiter option was found in the options dictionary.",
+            ):
+                optimiser(cost=cost, maxiter=5, options=dict(maxiter=10))
 
         if optimiser in [pybop.SciPyDifferentialEvolution]:
-            # Check and update population size
-            with pytest.raises(ValueError):
-                optim.set_population_size(-5)
-            optim.set_population_size(10)
-            assert optim.popsize == 10
-            optim.run(popsize=5)
-            assert optim.popsize == 5
+            # Update population size
+            optimiser(cost=cost, popsize=5)
 
             # Test invalid bounds
             with pytest.raises(ValueError):
-                optim.run(bounds=None)
+                optimiser(cost=cost, bounds=[(0, np.inf)])
             with pytest.raises(ValueError):
-                optim.run(bounds=[(0, np.inf)])
-            with pytest.raises(ValueError):
-                optim.run(bounds={"upper": [np.inf], "lower": [0.57]})
+                optimiser(cost=cost, bounds={"upper": [np.inf], "lower": [0.57]})
 
         else:
             # Check and update initial values
             assert optim.x0 == cost.x0
             x0_new = np.array([0.6])
-            optim.run(x0=x0_new)
+            optim = optimiser(cost=cost, x0=x0_new)
             assert optim.x0 == x0_new
 
         if optimiser in [pybop.SciPyMinimize]:
             # Check a method that uses gradient information
-            optim.run(method="L-BFGS-B", jac=True)
+            optimiser(cost=cost, method="L-BFGS-B", jac=True)
             with pytest.raises(
                 ValueError,
                 match="Expected the jac option to be either True, False or None.",
             ):
-                optim.run(jac="Invalid string")
-            optim.run(method="Nelder-Mead", jac=False)
+                optim = optimiser(cost=cost, jac="Invalid string")
+                optim.run()
 
     @pytest.mark.unit
     def test_single_parameter(self, cost):
@@ -236,18 +218,20 @@ class TestOptimisation:
         class RandomClass:
             pass
 
-        with pytest.raises(ValueError):
-            pybop.BasePintsOptimiser(cost=cost, pints_method=RandomClass)
+        with pytest.raises(
+            ValueError,
+            match="The pints_optimiser is not a recognised PINTS optimiser class.",
+        ):
+            pybop.BasePintsOptimiser(cost=cost, pints_optimiser=RandomClass)
 
-        optim = pybop.Optimisation(cost=cost)
         with pytest.raises(NotImplementedError):
-            optim.run()
+            pybop.Optimisation(cost=cost)
 
     @pytest.mark.unit
     def test_prior_sampling(self, cost):
         # Tests prior sampling
         for i in range(50):
-            optim = pybop.Optimisation(cost=cost)
+            optim = pybop.DefaultOptimiser(cost=cost)
 
             assert optim.x0 <= 0.62 and optim.x0 >= 0.58
 
@@ -284,13 +268,12 @@ class TestOptimisation:
     def test_infeasible_solutions(self, cost):
         # Test infeasible solutions
         for optimiser in [pybop.SciPyMinimize, pybop.GradientDescent]:
-            optim = optimiser(cost=cost, allow_infeasible_solutions=False)
-            optim.set_max_iterations(1)
+            optim = optimiser(cost=cost, allow_infeasible_solutions=False, maxiter=1)
             optim.run()
             assert optim._iterations == 1
 
     @pytest.mark.unit
     def test_unphysical_result(self, cost):
         # Trigger parameters not physically viable warning
-        optim = pybop.Optimisation(cost=cost)
+        optim = pybop.DefaultOptimiser(cost=cost)
         optim.check_optimal_parameters(np.array([2]))